Pass vectors instead of actors to the sorting functor. Run on client.
AActors* are not yet available at the point sorting is done so
to_actor() always returns null. However position vector is available
already, and that's what sorting needs anyways, so pass that to the
DelegateChain without bothering with Actor pointers.
Also make sure functors in DelegateChain are added and removed on
the client and not on the server.